refactor(guilds/DurationInput): move logic into its own hook
